Close Issue #25
The default REGEX for URL variables is found in this function: **tethys_apps.base.url_map:django_url_preprocessor()**

I expanded the default REGEX to include underscores (_) and periods (.) as follows:

DEFAULT_EXPRESSION = '[0-9A-Za-z-_.]+'

Note: dashes (-) were already included.

I also added a new optional parameter to the UrlMap object to allow for a custom REGEX to be set. The parameter is called 'regex' and it will accept a string or a list/tuple of strings to support different regex for urls with multiple parameters.